requirements.txt
slack_sdk
python-dotenv
tenacity
tqdm